pyproject.toml
[project]
name = "super-mega-data-gatherer"
version = "0.1.0"
description = "Gathers support data for the Super Mega Dashboard"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi[standard]>=0.129.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
]
